\name{countryref}
\alias{countryref}
\docType{data}
\title{
Country Centroids and Country Capitals
}
\description{
This is a data.frame with coordinates of country centroids and country capitals for the \code{\link{GeoClean}} function.  Coordinates are based on the Central Intelligence Agency World Factbook as provided at http://opengeocode.org/download/cow.php.
}
\usage{data("countryref")}
\format{
  A data frame with 249 observations on 7 variables.
 }
\source{
CENTRAL INTELLIGENCE AGENCY (2014) \emph{The World Factbook}, Washington, DC.

\url{http://opengeocode.org/download/cow.php}
}
\examples{
data(countryref)
}
\keyword{datasets}
