\name{institutions}
\alias{institutions}
\docType{data}
\title{
Geographic coordinates and addresses of biodiversity institutions.
}
\description{
A global gazetteer for biodiviersity institutions from various sources, including zoos, museums, botanical gardens, GBIF contributors, herbaria, university collections.
}
\usage{data("institutions")}
\format{
  A data frame with 5987 observations on 9 variables.
}

\source{
various
}

\examples{
data(institutions)
str(institutions)
}
\keyword{datasets}
