\name{WriteOut}
\alias{WriteOut}

\title{
Write spgeoOUT Objects to the Working Directory
}
\description{
Writes objects of the class \sQuote{spgeoOUT} (as generated by \code{SpGeoCod}) to the working directory. Output statistics as .txt files, maps and plots as .pdf and a nexus file for use in phylogenetic analyses.
}
\usage{
WriteOut(x, type = "all")
}

\arguments{
  \item{x}{
an object of the class spgeoOUT
}
  \item{type}{
a character string. Defines which results should be written to disk. One of "all", "BioGeoBEARS", "BayArea", "graphs", "maps", "nexus", "stats". Default is to \dQuote{all}.
}
}
\details{
There are different options for the \code{writetype} argument: \dQuote{all} = all summary tables, plots and maps; \dQuote{BioGeoBEARS} = a text file in the geography input format of BioGeoBEARS; \dQuote{BayArea} = two text files in the input format of BayArea; \dQuote{graphs} = summary graphs as pdf files; \dQuote{maps} = summary maps as .pdf file; \dQuote{nexus} = the species to area classification as nexus file; \dQuote{stats} = the summary tables as .txt files.
}
\value{
a set of files in the working directory
}

\note{
See \url{https://github.com/azizka/speciesgeocodeR/wiki} for more details and tutorials.
}

\seealso{
\code{\link{SpGeoCod}}
}
\examples{
\dontrun{
data(lemurs)
data(mdg_biomes)

outp <- SpGeoCodH(lemurs, mdg_biomes, areanames = 'name')
WriteOut(outp)
WriteOut(outp, writetype = "graphs")
WriteOut(outp, writetype = "statistics")
WriteOut(outp, writetype = "nexus")
}
}

\keyword{ Area classification }